+++ fhir-context/dstu2/index.d.ts
@@ -5,5 +5,8 @@ export const {
     choiceTypePaths,
     pathsDefinedElsewhere,
     type2Parent,
-    path2Type
+    path2Type,
+    resourcesWithUrlParam,
+    path2TypeWithoutElements,
+    path2RefType
 }: Model;