+++ src/fhirpath.js
@@ -31,6 +31,9 @@
 const {version} = require('../package.json');
 const parser = require("./parser");
 const util = require("./utilities");
+// Cannot use util.hasOwnProperty directly because it triggers the error:
+// "Do not access Object.prototype method 'hasOwnProperty' from target object"
+const { hasOwnProperty } = util;
 require("./polyfill");
 const constants = require('./constants');
 
@@ -66,6 +69,7 @@ const Factory = require('./factory');
 //   calling function if one of params is  empty return empty
 
 engine.invocationTable = {
+  resolve:      {fn: additional.resolveFn},
   memberOf:     {fn: additional.memberOf, arity: { 1: ['Any']} },
   empty:        {fn: existence.emptyFn},
   not:          {fn: existence.notFn},
@@ -558,7 +562,7 @@ function makeParam(ctx, parentData, type, param) {
 function doInvoke(ctx, fnName, data, rawParams){
   var invoc =
     ctx.userInvocationTable
-    && Object.prototype.hasOwnProperty.call(ctx.userInvocationTable, fnName)
+    && hasOwnProperty(ctx.userInvocationTable, fnName)
     && ctx.userInvocationTable?.[fnName]
     || engine.invocationTable[fnName]
     || data.length === 1 && data[0]?.invocationTable?.[fnName];
@@ -779,6 +783,8 @@ function parse(path) {
  * @param {string} [options.terminologyUrl] - a URL that points to a FHIR
  *   RESTful API that is used to create %terminologies that implements
  *   the Terminology Service API.
+ * @param {string} [options.fhirServerUrl] - a URL that points to a FHIR
+ *   RESTful API that is used to query resources when using `resolve()`.
  * @param {AbortSignal} [options.signal] - an AbortSignal object that allows you
  *   to abort the asynchronous FHIRPath expression evaluation.
  */
@@ -819,6 +825,9 @@ function applyParsedPath(resource, parsedPath, envVars, model, options) {
   }
   if (options.terminologyUrl) {
     ctx.processedVars.terminologies = new Terminologies(options.terminologyUrl);
+  }
+  if (options.fhirServerUrl) {
+    ctx.processedVars.fhirServerUrl = options.fhirServerUrl;
   }
   ctx.processedVars.factory = Factory;
   if (options.signal) {
@@ -947,6 +956,8 @@ function resolveInternalTypes(val) {
  * @param {string} [options.terminologyUrl] - a URL that points to a FHIR
  *   RESTful API that is used to create %terminologies that implements
  *   the Terminology Service API.
+ * @param {string} [options.fhirServerUrl] - a URL that points to a FHIR
+ *   RESTful API that is used to query resources when using `resolve()`.
  * @param {AbortSignal} [options.signal] - an AbortSignal object that allows you
  *   to abort the asynchronous FHIRPath expression evaluation.
  */
@@ -979,6 +990,8 @@ function evaluate(fhirData, path, envVars, model, options) {
  * @param {string} [options.terminologyUrl] - a URL that points to a FHIR
  *   RESTful API that is used to create %terminologies that implements
  *   the Terminology Service API.
+ * @param {string} [options.fhirServerUrl] - a URL that points to a FHIR
+ *   RESTful API that is used to query resources when using `resolve()`.
  * @param {AbortSignal} [options.signal] - an AbortSignal object that allows you
  *   to abort the asynchronous FHIRPath expression evaluation. Passing a signal
  *   to compile() whose result is used more than once will cause abortion