+++ test/async-functions.test.js
@@ -1,10 +1,13 @@
 const fhirpath = require('../src/fhirpath');
+const modelDSTU2 = require('../fhir-context/dstu2');
+const modelSTU3 = require('../fhir-context/stu3');
 const modelR4 = require('../fhir-context/r4');
 const modelR5 = require('../fhir-context/r5');
+const _ = require("lodash");
 const emptyResource = {};
-const observationResource = require('./resources/observation-example.json');
-const administrativeGenderVS = require('./resources/administrative-gender-valueset.json');
-const observationCategoryVS = require('./resources/observation-category-valuset.json');
+const observationResource = require('./resources/r4/observation-example.json');
+const administrativeGenderVS = require('./resources/r4/administrative-gender-valueset.json');
+const observationCategoryVS = require('./resources/r4/observation-category-valuset.json');
 const {mockRestore, mockFetchResults} = require('./mock-fetch-results');
 
 
@@ -409,6 +412,52 @@ describe('Async functions', () => {
       expect(result).toEqual([]);
     });
 
+    it('should return an empty collection when a code does not have a value', () => {
+      mockFetchResults([
+        ['ValueSet?url=http%3A%2F%2Fhl7.org%2Ffhir%2FValueSet%2Fobservation-vitalsignresult', {
+          "resourceType": "Bundle",
+          "entry": [{
+            "resource": {
+              "resourceType": "ValueSet",
+              "url": "http://hl7.org/fhir/ValueSet/observation-vitalsignresult",
+              "compose": {
+                "include": [{
+                  "system": "http://loinc.org",
+                  "concept": [
+                    {"code": "85353-1"}, {"code": "9279-1"}, {"code": "8867-4"},
+                    {"code": "2708-6"}, {"code": "8310-5"}, {"code": "8302-2"},
+                    {"code": "9843-4"}, {"code": "29463-7"},
+                    {"code": "39156-5"}, {"code": "85354-9"},
+                    {"code": "8480-6"}, {"code": "8462-4"}, {"code": "8478-0"}
+                  ]
+                }]
+              }
+            }
+          }]
+        }]
+      ]);
+
+      let result = fhirpath.evaluate(
+        {
+          "resourceType": "Observation",
+          "code": {
+            "coding":
+              [{
+                "_code": {
+                  "id": "someCodeId"
+                }
+              }]
+          }
+        },
+        "%terminologies.validateVS('http://hl7.org/fhir/ValueSet/observation-vitalsignresult', Observation.code.coding.code)",
+        {},
+        modelR4,
+        { async: true, terminologyUrl: "https://lforms-fhir.nlm.nih.gov/baseR4" }
+      );
+
+      expect(result).toEqual([]);
+    });
+
 
     it('should throw an error when the async function is not allowed', () => {
       let result = () => fhirpath.evaluate(
@@ -667,7 +716,11 @@ describe('Async functions', () => {
       );
       expect(result instanceof Promise).toBe(true);
       result.then((r) => {
-        expect(r).toEqual([{"code": "29463-7", "display": "Body Weight", "system": "http://loinc.org"}]);
+        expect(r).toEqual([expect.objectContaining({
+          "code": "29463-7",
+          "display": "Body Weight",
+          "system": "http://loinc.org"
+        })]);
         done();
       })
     });
@@ -675,7 +728,11 @@ describe('Async functions', () => {
 
     it('should work with CodeableConcept when async functions are enabled', (done) => {
       mockFetchResults([
-        [/ValueSet\/\$validate-code/, {
+        [{
+          url: '/ValueSet/$validate-code',
+          body: '"system":"http://loinc.org"',
+          method: 'POST'
+        }, {
           "resourceType": "Parameters",
           "parameter": [
             {
@@ -698,6 +755,79 @@ describe('Async functions', () => {
         done();
       })
     });
+
+
+    it('should use GET for CodeableConcept with a single Coding', (done) => {
+      mockFetchResults([
+        [{
+          url: /ValueSet\/\$validate-code.*&system=system1/,
+          method: 'GET'
+        }, {
+          "resourceType": "Parameters",
+          "parameter": [
+            {
+              "name": "result",
+              "valueBoolean": true
+            }
+          ]
+        }]
+      ]);
+      let result = fhirpath.evaluate(
+        observationResource,
+        "%factory.CodeableConcept(%factory.Coding('system1', '1')).memberOf('http://some-valueset')",
+        {},
+        modelR4,
+        { async: true, terminologyUrl: "https://lforms-fhir.nlm.nih.gov/baseR4" }
+      );
+      expect(result instanceof Promise).toBe(true);
+      result.then((r) => {
+        expect(r).toEqual([true]);
+        done();
+      })
+    });
+
+    it('should work when Coding in a CodeableConcept has no system and code', () => {
+      mockFetchResults([]);
+      let result = fhirpath.evaluate(
+        {},
+        "%factory.CodeableConcept(%factory.Coding('', '', 'some display text')).memberOf('http://some-valueset')",
+        {},
+        modelR4,
+        { async: true, terminologyUrl: "https://lforms-fhir.nlm.nih.gov/baseR4" }
+      );
+      expect(result).toEqual([]);
+    });
+
+
+    it('should work when no Coding in a CodeableConcept', (done) => {
+      mockFetchResults([
+        [{
+          url: '/ValueSet/$validate-code',
+          body: '"valueUri":"http://some-valueset"',
+          method: 'POST'
+        }, null, {
+          "resourceType": "OperationOutcome",
+          "issue": [ {
+            "severity": "error",
+            "code": "processing",
+            "diagnostics": "HAPI-0899: No code, coding, or codeableConcept provided to validate"
+          } ]
+        }]
+      ]);
+      let result = fhirpath.evaluate(
+        {},
+        "%factory.CodeableConcept(Observation.code.coding).memberOf('http://some-valueset')",
+        {},
+        modelR4,
+        { async: true, terminologyUrl: "https://lforms-fhir.nlm.nih.gov/baseR4" }
+      );
+      expect(result instanceof Promise).toBe(true);
+      result.then((r) => {
+        expect(r).toEqual([]);
+        done();
+      });
+    });
+
 
 
     it('should work with "code" when async functions are enabled', (done) => {
@@ -902,6 +1032,230 @@ describe('Async functions', () => {
   });
 
 
+  /**
+   * Test suite for the 'resolve' functionality in FHIRPath evaluation.
+   * This suite verifies that various forms of FHIR references (relative/absolute URLs, canonical, uri, etc.)
+   * are correctly resolved to their corresponding resources using asynchronous evaluation.
+   */
+  describe('resolve', () => {
+    // Load test resources and set up test data for MedicationDispense and Patient resources.
+    const medicationDispenseResource106 = require('./resources/r4/medicationdispense-med-106-0.json');
+    const medicationDispenseResource107 = _.cloneDeep(medicationDispenseResource106);
+    medicationDispenseResource107.id = 'med-107-0';
+    medicationDispenseResource107.subject = {
+      "reference": "https://some-fhir-server/Patient/pat-107"
+    };
+    const medicationDispenseResource108 = _.cloneDeep(medicationDispenseResource106);
+    medicationDispenseResource108.id = 'med-108-0';
+    medicationDispenseResource108.subject = {
+      "reference": "https://some-fhir-server/Patient/pat-108"
+    };
+    const patientResource106 = require('./resources/r4/patient-pat-106-0.json');
+    const patientResource107 = _.cloneDeep(patientResource106);
+    patientResource107.id = 'pat-107';
+    patientResource107.deceasedDateTime = '2129';
+    const patientResource108 = _.cloneDeep(patientResource106);
+    patientResource108.id = 'pat-108';
+    patientResource108.deceasedDateTime = '2130';
+
+    // ValueSet bundle resource for canonical URL resolution tests.
+    const valueSetResource = {
+      "resourceType": "Bundle",
+      "entry": [{
+        "resource": {
+          "resourceType": "ValueSet",
+          "url": "http://some-canonical-value-set-url"
+        }
+      }]
+    };
+
+    // Questionnaire resource with contained/nested questionnaires for fragment resolution tests.
+    const questionnaireWithContainedQ = {
+      "resourceType": "Questionnaire",
+      "url": "http://some-canonical-questionnaire-url",
+      "version": "1.0",
+      "derivedFrom": "#childQuestionnaire",
+      "contained": [{
+        "resourceType": "Questionnaire",
+        "id": "childQuestionnaire",
+        "derivedFrom": "#childOfChildQuestionnaire",
+        "contained": [{
+          "resourceType": "Questionnaire",
+          "id": "childOfChildQuestionnaire"
+        }]
+      }]
+    };
+
+    /**
+     * Sets up mock fetch results before each test to simulate FHIR server responses.
+     */
+    beforeEach(() => {
+      mockFetchResults([
+        ['https://some-fhir-server/MedicationDispense/med-106-0', medicationDispenseResource106],
+        ['https://some-fhir-server/MedicationDispense/med-107-0', medicationDispenseResource107],
+        ['https://some-fhir-server/Patient/pat-106', patientResource106],
+        ['https://some-fhir-server/Patient/pat-107', patientResource107],
+        ['https://some-fhir-server/Patient/pat-108', null, {
+          "resourceType": "OperationOutcome",
+          "issue": [ {
+            "severity": "error",
+            "code": "processing",
+            "diagnostics": "Resource Patient/pat-108 is not known"
+          } ]
+        }],
+        // This impossible request should not be called, but is included to
+        // catch it in case it is called by mistake:
+        ['https://some-fhir-server/Patient?url=https%3A%2F%2Fsome-fhir-server%2FPatient%2Fpat-108', {
+          "resourceType": "Bundle",
+          "entry": [{
+            "resource": patientResource108
+          }]
+        }],
+        [/https:\/\/some-fhir-server\/ValueSet\?url=http%3A%2F%2Fsome-canonical-value-set-url$/, valueSetResource],
+        [/https:\/\/some-fhir-server\/ValueSet\?url=http%3A%2F%2Fsome-canonical-value-set-url&version=1.0$/, valueSetResource],
+        [/https:\/\/some-fhir-server\/Questionnaire\?url=http%3A%2F%2Fsome-canonical-questionnaire-url&version=2.0$/, {
+          "resourceType": "Bundle",
+          "entry": [{
+            "resource": questionnaireWithContainedQ
+          }]
+        }],
+        ['http://some-canonical-url', null, {
+          "resourceType": "OperationOutcome",
+          "issue": [{
+            "severity": "error",
+            "code": "processing",
+            "diagnostics": "Additional diagnostic information about the issue."
+          }]
+        }]
+      ]);
+    });
+
+    /**
+     * Parameterized tests for various resolve scenarios.
+     * Each test case checks that the FHIRPath expression resolves the reference as expected.
+     */
+    [
+      // [models, description, resource, expression, expectedResult]
+      [
+        [modelDSTU2, modelSTU3, modelR4],
+        'String with relative URL',
+        {},
+        '\'MedicationDispense/med-106-0\'.resolve().medication.coding.where(system=\'357\').code',
+        ['00168022138']
+      ],
+      [
+        [modelDSTU2, modelSTU3, modelR4],
+        'String with absolute URL',
+        {},
+        '\'https://some-fhir-server/MedicationDispense/med-107-0\'.resolve().medication.coding.where(system=\'357\').code',
+        ['00168022138']
+      ],
+      [
+        [modelSTU3, modelR4, modelR5],
+        'Reference with relative URL',
+        medicationDispenseResource106,
+        'MedicationDispense.subject.resolve().deceased.where($this is dateTime)',
+        ['2128']
+      ],
+      [
+        [modelSTU3, modelR4, modelR5],
+        'Reference with absolute URL',
+        medicationDispenseResource107,
+        'MedicationDispense.subject.resolve().deceased.where($this is dateTime)',
+        ['2129']
+      ],
+      [
+        [modelSTU3, modelR4, modelR5],
+        'Reference with absolute URL',
+        medicationDispenseResource108,
+        'MedicationDispense.subject.resolve().deceased.where($this is dateTime)',
+        []
+      ],
+      [
+        [modelDSTU2, modelSTU3, modelR4],
+        'uri with relative URL',
+        {},
+        '%factory.uri(\'MedicationDispense/med-106-0\').resolve().medication.coding.where(system=\'357\').code',
+        ['00168022138']
+      ],
+      [
+        [modelDSTU2, modelSTU3, modelR4],
+        'uri with absolute URL',
+        {},
+        '%factory.uri(\'https://some-fhir-server/MedicationDispense/med-107-0\').resolve().medication.coding.where(system=\'357\').code',
+        ['00168022138']
+      ],
+      [
+        [modelR4, modelR5],
+        'canonical for a synthetic node as an empty collection',
+        {},
+        '%factory.canonical(\'http://some-canonical-url\').resolve()',
+        []
+      ],
+      [
+        [modelR4, modelR5],
+        'canonical for a node that resolves to a resource',
+        {
+          "resourceType": "CodeSystem",
+          "valueSet": "http://some-canonical-value-set-url",
+        },
+        'CodeSystem.valueSet.resolve() is ValueSet',
+        [true]
+      ],
+      [
+        [modelR4, modelR5],
+        'canonical with version and fragment',
+        {
+          "resourceType": "QuestionnaireResponse",
+          "questionnaire": "http://some-canonical-questionnaire-url|2.0#childQuestionnaire",
+        },
+        'QuestionnaireResponse.questionnaire.resolve().where($this is Questionnaire).id=\'childQuestionnaire\'',
+        [true]
+      ],
+      [
+        [modelR4, modelR5],
+        'canonical URL with a fragment only',
+        questionnaireWithContainedQ,
+        'Questionnaire.derivedFrom.resolve().where($this is Questionnaire).id=\'childQuestionnaire\'',
+        [true]
+      ],
+      [
+        [modelR4, modelR5],
+        'canonical URL with a fragment only for a nested resource that contains another resource',
+        questionnaireWithContainedQ,
+        'Questionnaire.derivedFrom.resolve().derivedFrom.resolve().where($this is Questionnaire).id=\'childOfChildQuestionnaire\'',
+        [true]
+      ]
+    ].forEach((testData) => {
+      // Extract models and test parameters for each test case.
+      const [models, dataType, resource, expression, res] = testData;
+
+      // Run the test for each FHIR model version.
+      models.forEach(model => {
+        it(`should resolve ${dataType} (${model.version.toUpperCase()})`, (done) => {
+          let result = fhirpath.evaluate(
+            resource,
+            `${expression}`,
+            {},
+            model,
+            {
+              async: true,
+              fhirServerUrl: "https://some-fhir-server"
+            }
+          );
+          expect(result instanceof Promise).toBe(true);
+          result.then((r) => {
+            expect(r).toEqual(res);
+            done();
+          });
+        });
+
+      });
+    });
+
+  });
+
+
   it('should be a conversion of the result to a Promise when option async is set to "always"', (done) => {
     let result = fhirpath.evaluate(
       observationResource,
@@ -912,7 +1266,11 @@ describe('Async functions', () => {
     );
     expect(result instanceof Promise).toBe(true);
     result.then((r) => {
-      expect(r).toEqual([{"code": "29463-7", "display": "Body Weight", "system": "http://loinc.org"}]);
+      expect(r).toEqual([expect.objectContaining({
+        "code": "29463-7",
+        "display": "Body Weight",
+        "system": "http://loinc.org"
+      })]);
       done();
     })
   });
@@ -927,7 +1285,11 @@ describe('Async functions', () => {
       {}
     );
     expect(result instanceof Promise).toBe(false);
-    expect(result).toEqual([{"code": "29463-7", "display": "Body Weight", "system": "http://loinc.org"}]);
+    expect(result).toEqual([expect.objectContaining({
+      "code": "29463-7",
+      "display": "Body Weight",
+      "system": "http://loinc.org"
+    })]);
   });