+++ test/benchmark/runner.js
@@ -4,8 +4,8 @@
  */
 const _ = require('lodash');
 const current_fhirpath = require('../../src/fhirpath');
-const minimumDataset = require('../resources/Minimum-Data-Set---version-3.0.R4.json');
-const patientExample = require('../resources/patient-example.json');
+const minimumDataset = require('../resources/r4/Minimum-Data-Set---version-3.0.R4.json');
+const patientExample = require('../resources/r4/patient-example.json');
 const current_r4_model = require('../../fhir-context/r4');
 const benny = require('benny');
 const open = require('open');
@@ -81,6 +81,11 @@ async function run(filename, options) {
         benny.cycle(),
         benny.complete(),
         benny.configure({
+          cases: {
+            initCount: 5,
+            maxTime: 15,
+            minSamples: 100,
+          },
           minDisplayPrecision: 2
         }),
         benny.save({ file: suite.filename, folder: __dirname + '/results', version: currentVersion }),