+++ CHANGELOG.md
@@ -3,6 +3,34 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [4.5.1] - 2025-07-09
+### Added
+- tests for the R5 model.
+### Changed
+- Updated the "%terminologies.validateVS()" method and "memberOf()" function to
+  use GET for CodeableConcept with a single Coding when using a ValueSet
+  reference instead of an actual ValueSet.
+### Fixed
+- updated tests for the R4 model.
+
+## [4.5.0] - 2025-07-01
+### Added
+- additional function `resolve()`.
+### Fixed
+- added descriptions for nested nodes that were missing in the DSTU2 model.
+
+## [4.4.1] - 2025-06-18
+### Fixed
+- Issues with the `children()` and `descendants()` functions, which returned
+  incorrect results:
+  - `resourceType` is a virtual property and should not be in the results,
+  - the object that contains `id` and `extension` on primitive elements should
+    not be in the results,
+  - children of primitives are not returned (when directly queries e.g.
+    birthDate.children()).
+- Internal tracking of the full property name of a node in a resource.
+  (Fixed `fullPropertyName()` method of the `ResourceNode` class.)
+
 ## [4.4.0] - 2025-06-04
 ### Added
 - Button to copy a shareable link to the current expression and its parameters